Вынес события в отдельный модуль и перенес функции создания окон в utils
