+++ js/editForm.js
@@ -1,3 +1,4 @@
+import {createSuccessAlert, createErrorAlert} from './util.js';
 import {sendData} from './api.js';
 import {onChangeUploadInputSetDefaultEffect} from './imageEffects.js';
 import {onChangeUploadInputSetDefaultScale} from './imageScale.js';
@@ -132,64 +133,6 @@ pristine.addValidator(textDescription, (value) => {
   return true;
 }, 'длина комментария больше 140 символов');
 
-const editFormAlertOpen = new Event('editFormAlertOpen');
-const editFormAlertClose = new Event('editFormAlertClose');
-document.addEventListener('editFormAlertOpen', () => {
-  document.removeEventListener('keydown', onEscapeEditFormClose);
-});
-document.addEventListener('editFormAlertClose', () => {
-  document.addEventListener('keydown', onEscapeEditFormClose);
-});
-const createSuccessAlert = () => {
-  const templateSuccessAlert = document.querySelector('#success').content.querySelector('.success');
-  const elementSuccessAlert = templateSuccessAlert.cloneNode(true);
-  const closeAlertButton = elementSuccessAlert.querySelector('.success__button');
-  const onEscAlertClose = (evt) => {
-    if(evt.key === 'Escape') {
-      elementSuccessAlert.remove();
-      document.removeEventListener('keydown', onEscAlertClose);
-    }
-  };
-  closeAlertButton.addEventListener('click', () => {
-    elementSuccessAlert.remove();
-    document.removeEventListener('keydown', onEscAlertClose);
-  });
-  elementSuccessAlert.addEventListener('click', (evt) => {
-    if(!evt.target.closest('.success__inner')){
-      elementSuccessAlert.remove();
-      document.removeEventListener('keydown', onEscAlertClose);
-    }
-  });
-  document.addEventListener('keydown', onEscAlertClose);
-  document.body.appendChild(elementSuccessAlert);
-};
-const createErrorAlert = () => {
-  const templateErrorAlert = document.querySelector('#error').content.querySelector('.error');
-  const elementErrorAlert = templateErrorAlert.cloneNode(true);
-  const closeAlertButton = elementErrorAlert.querySelector('.error__button');
-  document.dispatchEvent(editFormAlertOpen);
-  const onEscAlertClose = (evt) => {
-    if(evt.key === 'Escape') {
-      elementErrorAlert.remove();
-      document.removeEventListener('keydown', onEscAlertClose);
-      document.dispatchEvent(editFormAlertClose);
-    }
-  };
-  closeAlertButton.addEventListener('click', () => {
-    elementErrorAlert.remove();
-    document.removeEventListener('keydown', onEscAlertClose);
-    document.dispatchEvent(editFormAlertClose);
-  });
-  elementErrorAlert.addEventListener('click', (evt) => {
-    if(!evt.target.closest('.error__inner')){
-      elementErrorAlert.remove();
-      document.removeEventListener('keydown', onEscAlertClose);
-      document.dispatchEvent(editFormAlertClose);
-    }
-  });
-  document.addEventListener('keydown', onEscAlertClose);
-  document.body.appendChild(elementErrorAlert);
-};
 const submitButton = document.querySelector('.img-upload__submit');
 const onSuccessSend = () => {
   closeEditForm();
@@ -211,3 +154,5 @@ imageUploadForm.addEventListener('submit', (evt) => {
     sendData(PICTURES_CREATE_URL, formData, onSuccessSend, onErrorSend);
   }
 });
+
+export {onEscapeEditFormClose};
